requirements.txt
PyYAML==5.1.2
requests
numpy==1.17.2
pandas==0.25.1
shapely==1.6.4.post2
plotly==4.1.1
scikit-learn==0.21.3
# local package
-e .

# external requirements
click
Sphinx
coverage
# awscli
flake8
python-dotenv==0.10.3